Added REAMDE.txt and .jar

+++ src/src/FtpRequest.java
@@ -121,20 +121,23 @@ public class FtpRequest extends Thread {
 	 * @throws IOException
 	 */
 	public void processRequest() throws IOException {
+		String request = null;
 		String cmd = null;
 		String param = null;
 
 		do {
+			request = null;
 			// Reception d'une requête sur l'entrée du socket
-			String request = this.input.readLine();
+			// On boucle pour éviter la reception de requête null qui peut arriver
+			while (request == null) {
+				request = this.input.readLine();
+			}
 
 			System.out.println("Request: " + request);
 
-			if (request != null) {
-				int index = request.indexOf(" ");
-				cmd = index != -1 ? request.substring(0, index) : request;
-				param = index != -1 ? request.substring(index + 1) : null;
-			}
+			int index = request.indexOf(" ");
+			cmd = index != -1 ? request.substring(0, index) : request;
+			param = index != -1 ? request.substring(index + 1) : null;
 
 			// Vérification de la validité de la commande : Elle doit exister et
 			// avoir un nombre de paramêtre correcte. L'utilisateur doit aussi
@@ -474,7 +477,7 @@ public class FtpRequest extends Thread {
 	 * 
 	 * @param filename
 	 *            le nom du fichier envoyé au client
-	 * @return
+	 * @return vrai si la commande se déroule bien, faux sinon
 	 */
 	public boolean processRETR(String filename) {
 		if (socketDonnees == null) {
@@ -484,21 +487,30 @@ public class FtpRequest extends Thread {
 		}
 
 		try {
-			String msg = String.valueOf(Constantes.CODE_LIST);
-			this.sendMessage(msg);
 			String path = this.repertoire.toPath().toAbsolutePath().toString() + "/" + filename;
-			Path target = Paths.get(path);
-
-			OutputStream out = this.socketDonnees.getOutputStream();
-
-			Files.copy(target, out);
-
-			socketDonnees.close();
-
-			if (this.serveur != null) {
-				this.serveur.close();
-				this.serveur = null;
+			
+			File f = new File(path);
+			if(f.exists()) { 
+				String msg = String.valueOf(Constantes.CODE_LIST);
+				this.sendMessage(msg);
+				
+				Path target = Paths.get(path);
+				OutputStream out = this.socketDonnees.getOutputStream();
+				
+				Files.copy(target, out);
+				
+				socketDonnees.close();
+				
+				if (this.serveur != null) {
+					this.serveur.close();
+					this.serveur = null;
+				}
+			} else {
+				String msg = Constantes.CODE_TRANSFERT_ECHOUE + " " + Constantes.MSG_FICHIER_INEXISTANT;
+				this.sendMessage(msg);
+				return false;
 			}
+
 		} catch (Exception e) {
 			String msg = Constantes.CODE_TRANSFERT_ECHOUE + " " + Constantes.MSG_TRANSFERT_ECHOUE;
 			this.sendMessage(msg);
@@ -514,7 +526,7 @@ public class FtpRequest extends Thread {
 	 * 
 	 * @param filename
 	 *            le nom du fichier copié depuis le cient
-	 * @return
+	 * @return vrai si la commande se déroule bien, faux sinon
 	 */
 	public boolean processSTOR(String filename) {
 		if (socketDonnees == null) {
@@ -564,6 +576,7 @@ public class FtpRequest extends Thread {
 	 * @param chemin
 	 *            le chemin du dossier vers lequel on souhaite naviguer
 	 * @throws IOException
+	 * @return le code de retour de la commande
 	 */
 	public int processCWD(String chemin) throws IOException {
 		Path path = Paths.get(this.repertoire.getPath() + "/" + chemin);
@@ -584,6 +597,7 @@ public class FtpRequest extends Thread {
 	 * Méthode éxécutant la commande CDUP
 	 * 
 	 * @throws IOException
+	 * @return le code de retour de la commande
 	 */
 	public int processCDUP() throws IOException {
 		Path path;
